refactor(BetterTerminal): rewrite some of BetterTerminal for myself to better understand what I was doing fix(BetterTerminal): fix clicks being delayed 1-2 more ticks than expected fix(BetterTerminal): fix terminal being reset in rubix and panes when misclick

+++ src/main/kotlin/yqloss/yqlossclientmixinkt/module/betterterminal/BetterTerminalOptions.kt
@@ -26,8 +26,8 @@ interface BetterTerminalOptions : YCModuleOptions {
     val preventFail: Boolean
     val preventMisclick: Boolean
     val reloadOnMismatch: Boolean
-    val clickDelayFrom: Int
-    val clickDelayUntil: Int
+    val clickDelayFrom: Double
+    val clickDelayUntil: Double
     val orderEnabled: Boolean
     val orderShowNumber: Boolean
     val orderShowClickedNumber: Boolean
